+++ scout/server/blueprints/variants/views.py
@@ -12,7 +12,7 @@ from scout.constants import ACMG_MAP
 from scout.server.extensions import store, mail, loqusdb
 from scout.server.utils import templated, institute_and_case, public_endpoint
 from scout.utils.acmg import get_acmg
-from scout.parse.clinvar import get_submission_variants, get_submission_header, get_submission_lines, create_clinvar_submission_dict, extract_submission_csv_lines
+from scout.parse.clinvar import set_submission_objects
 from . import controllers
 from .forms import FiltersForm, SvFiltersForm, StrFiltersForm
 
@@ -248,125 +248,17 @@ def clinvar(institute_id, case_name, variant_id):
     data = controllers.clinvar_export(store, institute_id, case_name, variant_id)
     if request.method == 'GET':
         return data
-    else:
+    else: #POST
         form_dict = request.form.to_dict()
-        variants_to_submit = get_submission_variants(form_dict) # list of variants to be submitted
-        variant_header = get_submission_header(form_dict, variants_to_submit, 'variants')
-        variant_lines = get_submission_lines(form_dict, variants_to_submit, variant_header)
-        casedata_header = get_submission_header(form_dict, variants_to_submit, 'casedata')
-        casedata_lines = get_submission_lines(form_dict, variants_to_submit, casedata_header)
-        variant_types = {}
-        for var in variants_to_submit:
-            variant_types[var] = form_dict['variant-type_'+str(var)]
+        submission_objects = set_submission_objects(form_dict) # A tuple of submission objects (variants and casedata objects)
 
-        # create clinvar submission session object:
-        session['clinvar_submission'] = create_clinvar_submission_dict(variant_header, variant_lines, casedata_header, casedata_lines, variant_types)
-        data.update({'variant_header': variant_header, 'variant_lines': variant_lines, 'casedata_header': casedata_header, 'casedata_lines': casedata_lines, 'form':request.form,})
-        return data
+        # Add submission data to an open clinvar submission object,
+        # or create a new if no open submission is found in database
+        open_submission = store.get_open_clinvar_submission(current_user.email, institute_id)
+        updated_submission = store.add_to_submission(open_submission['_id'], submission_objects)
 
-
-@variants_bp.route('/get_csv/', methods=['POST','GET'])
-def get_csv():
-    """Creates csv files (.Variant.csv or .CaseData.csv) to be used for submitting variants to clinVar."""
-    def generate(header, lines):
-        yield '"'+header+'"' + '\n'
-        for line in lines: # lines have already quoted fields
-            yield line + '\n'
-    if request.form.get('variants_button'):
-        header = request.form['vheader']
-        lines = request.form.getlist('variant')
-        filename = str(request.form.get('subm_id')) + '.Variant.csv'
-    else:
-        header = request.form['cdheader']
-        lines = request.form.getlist('case')
-        filename = str(request.form.get('subm_id')) + '.CaseData.csv'
-
-    headers = Headers()
-    headers.add('Content-Disposition','attachment', filename=filename)
-    return Response(generate(header, lines), mimetype='text/csv', headers=headers)
-
-
-@variants_bp.route('/<institute_id>/<case_name>/<variant_id>', methods=['POST'])
-def save_clinvar_submission(institute_id, case_name, variant_id):
-    """Saves variants submitted to clinVar to database and redirects to variants page"""
-    # clinvar submission form exists in this session, save it to mongo db:
-    if session.get('clinvar_submission') and request.form.get('subm_id'):
-        for variant_submission in session.get('clinvar_submission'):
-            variant_submission['clinvar_submission'] = request.form.get('subm_id')
-        inserted = store.add_clinvar_submission(session.get('clinvar_submission'),current_user.email, institute_id, case_name)
-
-        if inserted == 0:
-            flash('Clinvar submission id '+str(request.form.get('subm_id'))+' already exists in database!', 'danger' )
-        elif inserted == -1:
-            flash('One of more variants your are trying to save is already present in a previous clinvar submission!', 'danger')
-        else:
-            flash('variants were saved into clinvar submissions database collection', 'success')
-
-        return redirect(url_for('.variant', institute_id=institute_id, case_name=case_name,
-                            variant_id=variant_id))
-
-    else: # redirect to variant's page with error message:
-        if session.get('clinvar_submission'):
-            flash("didn't receive a valid clinvar submission id from the previous form", 'danger')
-        else:
-            flash('a session object named "clinvar_submission" could not be found!', 'danger')
-        return redirect(url_for('.variant', institute_id=institute_id, case_name=case_name,
-                            variant_id=variant_id))
-
-
-@variants_bp.route('/<institute_id>/<case_name>/<variant_id>/update_clinvar/<submission_id>', methods=['POST', 'GET'])
-@templated('variants/clinvar_update.html')
-def update_clinvar_submission(institute_id, case_name, variant_id, submission_id):
-    """Update/Removes a clinvar submission for a variant or a group of variants"""
-    def generate(header, lines):
-        yield header + '\n'
-        for line in lines:
-            yield line + '\n'
-
-    data = controllers.get_clinvar_submission(store, institute_id, case_name, variant_id, submission_id)
-    if request.method == 'GET':
-        return data
-    elif request.form.get('variants_button') or request.form.get('cdata_button'):
-        variants_header, casedata_header, clinvar_lines, casedata_lines = extract_submission_csv_lines(data['clinvars'])
-
-        if request.form.get('variants_button'):
-            filename = str(submission_id) + '.Variant.csv'
-            header = variants_header
-            lines = clinvar_lines
-
-        elif request.form.get('cdata_button'):
-            filename = str(submission_id) + '.CaseData.csv'
-            header = casedata_header
-            lines = casedata_lines
-
-        headers = Headers()
-        headers.add('Content-Disposition','attachment', filename=filename)
-        return Response(generate(header, lines), mimetype='text/csv', headers=headers)
-
-    else:
-        if request.form.get('add_accession'):
-            updates=[] #a list of tuples
-            for fieldname, value in request.form.items():
-                if not value == 'submit':
-                    updates.append(store.add_clinvar_accession(fieldname.replace('clinvar_accession_',''), value))
-            if len(updates) == 0:
-                flash('no updates done', 'info')
-            else:
-                flash('Clinvar variation ID has been updated', 'success')
-            return redirect(url_for('.update_clinvar_submission', institute_id=institute_id, case_name=case_name,
-                                variant_id=variant_id, submission_id=submission_id))
-
-        elif request.form.get('delete_submission'):
-            deleted = store.delete_clinvar_submission(submission_id)
-
-            if deleted:
-                flash('{} variants submitted to clinvar with submission id {} deleted from the database!'.format(deleted, submission_id), 'success')
-            else:
-                flash("Couldn't find any clinvar variant with submission id to remove {}.".format(submission_id), 'info')
-
-            return redirect(url_for('.variant', institute_id=institute_id, case_name=case_name,
-                                variant_id=variant_id))
-
+        # Redirect to clinvar submissions handling page, and pass it the updated_submission_object
+        return redirect(url_for('cases.clinvar_submissions', institute_id=institute_id))
 
 
 @variants_bp.route('/<institute_id>/<case_name>/cancer/variants')